move data structures and analyst tools out of aws module
