+++ sppy/tools/s2n/aggregate_matrix.py
@@ -9,8 +9,8 @@ from sppy.aws.aws_tools import (
     download_from_s3, get_current_datadate_str, get_today_str,
     read_s3_parquet_to_pandas
 )
-from sppy.aws.sparse_matrix import SparseMatrix
-from sppy.aws.summary_matrix import SummaryMatrix
+from sppy.tools.s2n.sparse_matrix import SparseMatrix
+from sppy.tools.s2n.summary_matrix import SummaryMatrix
 from sppy.tools.util.logtools import Logger, logit
