+++ sppy/tools/s2n/spnet.py
